znaki

+++ 1tmr_1/pole_prostokata.py
@@ -1,9 +1,9 @@
 a=0
 b=0
 while a <= 0:
-    a = int(input('podaj pierwszy bok='))
+    a = float(input('podaj pierwszy bok='))
 while b <= 0:
-    b = int(input('podaj drugi bok='))
+    b = float(input('podaj drugi bok='))
 
 
 pole = a*b
